move the randomizer patch into the randomizer class
